enclave_driver: Create a new crate for the enclave driver
Move all driver bindings into a separate crate and make enclave_api
crate use them.
