+++ enclave_driver/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "enclave_driver"
+version = "0.1.0"
+authors = ["Paul-Sebastian Ungurenu <paulsebu@amazon.com>"]
+edition = "2018"
+
+[dependencies]
+nix = "0.15"
+
+[build-dependencies]
+bindgen = { version=">=0.54" }
