+++ enclave_driver/src/includes.rs
@@ -1,0 +1,2 @@
+#![allow(non_camel_case_types)]
+include!(concat!(env!("OUT_DIR"), "/driver_structs.rs"));
